Move Picture commands to main interpreter because they work in battle for RPG2k3 1.12
