+++ src/game_interpreter_map.cpp
@@ -164,12 +164,6 @@ bool Game_Interpreter_Map::ExecuteCommand() {
 			return CommandEnterExitVehicle(com);
 		case Cmd::PanScreen:
 			return CommandPanScreen(com);
-		case Cmd::ShowPicture:
-			return CommandShowPicture(com);
-		case Cmd::MovePicture:
-			return CommandMovePicture(com);
-		case Cmd::ErasePicture:
-			return CommandErasePicture(com);
 		case Cmd::ShowBattleAnimation:
 			return CommandShowBattleAnimation(com);
 		case Cmd::FlashSprite:
@@ -624,177 +618,6 @@ bool Game_Interpreter_Map::CommandPanScreen(RPG::EventCommand const& com) { // c
 
 	if (waiting_pan_screen)
 		wait_count = distance * (2 << (6 - speed));
-
-	return true;
-}
-
-// PicPointerPatch handling.
-// See http://cherrytree.at/cms/download/?did=19
-namespace PicPointerPatch {
-
-static void AdjustId(int& pic_id) {
-	if (pic_id > 10000) {
-		int new_id;
-		if (pic_id > 50000) {
-			new_id = Game_Variables[pic_id - 50000];
-		} else {
-			new_id = Game_Variables[pic_id - 10000];
-		}
-
-		if (new_id > 0) {
-			Output::Debug("PicPointer: ID %d replaced with ID %d", pic_id, new_id);
-			pic_id = new_id;
-		}
-	}
-}
-
-static void AdjustParams(Game_Picture::Params& params) {
-	if (params.magnify > 10000) {
-		int new_magnify = Game_Variables[params.magnify - 10000];
-		Output::Debug("PicPointer: Zoom %d replaced with %d", params.magnify, new_magnify);
-		params.magnify = new_magnify;
-	}
-
-	if (params.top_trans > 10000) {
-		int new_top_trans = Game_Variables[params.top_trans - 10000];
-		Output::Debug("PicPointer: Top transparency %d replaced with %d", params.top_trans, new_top_trans);
-		params.top_trans = new_top_trans;
-	}
-
-	if (params.bottom_trans > 10000) {
-		int new_bottom_trans = Game_Variables[params.bottom_trans - 10000];
-		Output::Debug("PicPointer: Bottom transparency %d replaced with %d", params.bottom_trans, new_bottom_trans);
-		params.bottom_trans = new_bottom_trans;
-	}
-}
-
-static void AdjustShowParams(int& pic_id, Game_Picture::ShowParams& params) {
-	// Adjust name
-	if (pic_id >= 50000) {
-		// Name substitution is pic_id + 1
-		int pic_num = Game_Variables[pic_id - 50000 + 1];
-
-		if (pic_num >= 0 && params.name.size() >= 4) {
-			// Replace last 4 characters with 0-padded pic_num
-			std::u32string u_pic_name = Utils::DecodeUTF32(params.name);
-			std::string new_pic_name = Utils::EncodeUTF(u_pic_name.substr(0, u_pic_name.size() - 4));
-			std::stringstream ss;
-			ss << new_pic_name << std::setfill('0') << std::setw(4) << pic_num;
-			new_pic_name = ss.str();
-
-			Output::Debug("PicPointer: File %s replaced with %s", params.name.c_str(), new_pic_name.c_str());
-			params.name = new_pic_name;
-		}
-	}
-
-	AdjustId(pic_id);
-	AdjustParams(params);
-}
-
-static void AdjustMoveParams(int& pic_id, Game_Picture::MoveParams& params) {
-	AdjustId(pic_id);
-	AdjustParams(params);
-
-	if (params.duration > 10000) {
-		int new_duration = Game_Variables[params.duration - 10000];
-		Output::Debug("PicPointer: Move duration %d replaced with %d", params.duration, new_duration);
-		params.duration = new_duration;
-	}
-}
-
-}
-
-bool Game_Interpreter_Map::CommandShowPicture(RPG::EventCommand const& com) { // code 11110
-	int pic_id = com.parameters[0];
-
-	Game_Picture::ShowParams params;
-	params.name = com.string;
-	params.position_x = ValueOrVariable(com.parameters[1], com.parameters[2]);
-	params.position_y = ValueOrVariable(com.parameters[1], com.parameters[3]);
-	params.fixed_to_map = com.parameters[4] > 0;
-	params.magnify = com.parameters[5];
-	params.transparency = com.parameters[7] > 0;
-	params.top_trans = com.parameters[6];
-	params.red = com.parameters[8];
-	params.green = com.parameters[9];
-	params.blue = com.parameters[10];
-	params.saturation = com.parameters[11];
-	params.effect_mode = com.parameters[12];
-	params.effect_power = com.parameters[13];
-
-	if (Player::IsRPG2k() || Player::IsRPG2k3E()) {
-		// RPG2k and RPG2k3 1.10 do not support this option
-		params.bottom_trans = params.top_trans;
-	} else {
-		// Corner case when 2k maps are used in 2k3 (pre-1.10) and don't contain this chunk
-		size_t param_size = com.parameters.size();
-		params.bottom_trans = param_size > 14 ? com.parameters[14] : params.top_trans;
-	}
-
-	PicPointerPatch::AdjustShowParams(pic_id, params);
-
-	// Sanitize input
-	params.magnify = std::max(0, std::min(params.magnify, 2000));
-	params.top_trans = std::max(0, std::min(params.top_trans, 100));
-	params.bottom_trans = std::max(0, std::min(params.bottom_trans, 100));
-
-	Game_Picture* picture = Main_Data::game_screen->GetPicture(pic_id);
-	picture->Show(params);
-
-	return true;
-}
-
-bool Game_Interpreter_Map::CommandMovePicture(RPG::EventCommand const& com) { // code 11120
-	int pic_id = com.parameters[0];
-
-	Game_Picture::MoveParams params;
-	params.position_x = ValueOrVariable(com.parameters[1], com.parameters[2]);
-	params.position_y = ValueOrVariable(com.parameters[1], com.parameters[3]);
-	params.magnify = com.parameters[5];
-	params.top_trans = com.parameters[6];
-	params.red = com.parameters[8];
-	params.green = com.parameters[9];
-	params.blue = com.parameters[10];
-	params.saturation = com.parameters[11];
-	params.effect_mode = com.parameters[12];
-	params.effect_power = com.parameters[13];
-	params.duration = com.parameters[14];
-
-	bool wait = com.parameters[15] != 0;
-
-	if (Player::IsRPG2k() || Player::IsRPG2k3E()) {
-		// RPG2k and RPG2k3 1.10 do not support this option
-		params.bottom_trans = params.top_trans;
-	} else {
-		// Corner case when 2k maps are used in 2k3 (pre-1.10) and don't contain this chunk
-		size_t param_size = com.parameters.size();
-		params.bottom_trans = param_size > 16 ? com.parameters[16] : params.top_trans;
-	}
-
-	PicPointerPatch::AdjustMoveParams(pic_id, params);
-
-	// Sanitize input
-	params.magnify = std::max(0, std::min(params.magnify, 2000));
-	params.top_trans = std::max(0, std::min(params.top_trans, 100));
-	params.bottom_trans = std::max(0, std::min(params.bottom_trans, 100));
-	params.duration = std::max(0, std::min(params.duration, 10000));
-
-	Game_Picture* picture = Main_Data::game_screen->GetPicture(pic_id);
-	picture->Move(params);
-
-	if (wait)
-		SetupWait(params.duration);
-
-	return true;
-}
-
-bool Game_Interpreter_Map::CommandErasePicture(RPG::EventCommand const& com) { // code 11130
-	int pic_id = com.parameters[0];
-
-	PicPointerPatch::AdjustId(pic_id);
-
-	Game_Picture* picture = Main_Data::game_screen->GetPicture(pic_id);
-	picture->Erase();
 
 	return true;
 }
